update README
In order to get this product running without null data errors I needed to create an environment variable with my Meetup API key. Meetup is moving away from API keys so I am not sure if the underlying plugin is going to be updated to address that. 

On the Meetup API site it says:
> Note: We are no longer supporting new API keys, and will soon be restricting access using your existing key. As an alternative, you can use OAuth2. Learn more.